Extract FileDirectoryPath
